后台代码

+++ src/main/java/com/healthy/vision/mappers/SysUserPOMapper.java
@@ -9,67 +9,67 @@ public interface SysUserPOMapper {
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   long countByExample(SysUserPOExample example);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int deleteByExample(SysUserPOExample example);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
-  int deleteByPrimaryKey(Long sysUserId);
+  int deleteByPrimaryKey(Integer sysUserId);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int insert(SysUserPO record);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int insertSelective(SysUserPO record);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   List<SysUserPO> selectByExample(SysUserPOExample example);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
-  SysUserPO selectByPrimaryKey(Long sysUserId);
+  SysUserPO selectByPrimaryKey(Integer sysUserId);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int updateByExampleSelective(@Param("record") SysUserPO record, @Param("example") SysUserPOExample example);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int updateByExample(@Param("record") SysUserPO record, @Param("example") SysUserPOExample example);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int updateByPrimaryKeySelective(SysUserPO record);
 
   /**
    * 只读. 
-   * @mbg.generated  2018-12-26 10:46:20
+   * @mbg.generated  2018-12-27 15:44:14
    */
   int updateByPrimaryKey(SysUserPO record);
 }
